+++ storybook/lib/main.dart
@@ -7,6 +7,7 @@ import 'package:glade_forms_storybook/generated/locale_loader.g.dart';
 import 'package:glade_forms_storybook/localization_addon_custom.dart';
 import 'package:glade_forms_storybook/usecases/complex_object_mapping_example.dart';
 import 'package:glade_forms_storybook/usecases/composed/composed_example.dart';
+import 'package:glade_forms_storybook/usecases/composed/composed_with_own_inputs_example.dart';
 import 'package:glade_forms_storybook/usecases/composed/nested_composed_example.dart';
 import 'package:glade_forms_storybook/usecases/dependencies/checkbox_dependency_change.dart';
 import 'package:glade_forms_storybook/usecases/metadata_descriptor_example.dart';
@@ -111,6 +112,10 @@ class App extends StatelessWidget {
             WidgetbookUseCase(
               name: 'Nested composed forms',
               builder: (context) => const NestedComposedExample(),
+            ),
+            WidgetbookUseCase(
+              name: "Composed form with composed model's own inputs",
+              builder: (context) => const ComposedWithOwnInputsExample(),
             ),
           ],
         ),